% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.correlation.R
\name{bootstrap.correlation}
\alias{bootstrap.correlation}
\title{Bootstrap Estimation of Correlation}
\usage{
bootstrap.correlation(n=2000, x , y, plot = FALSE)
}
\arguments{
\item{n}{number of bootstrap samples.}

\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{plot}{specifies whether a plot should be created or not}
}
\value{
Returns a list containing: the estimated correlation of the two given vectors, the bootstrap estimated correlation of the data and the 95 percent confidence interval.
}
\description{
This function takes as arguments the number of bootstrap replicates \code{n},
\code{x} and \code{y} which are two numerical vectors / data, then bootstrap from \code{x ,y}, and calculate the correlation of these data for each resampling n times.
If \code{plot} is chosen to be TRUE, then the function plots the histogram, as well as the 95% confidence interval for the
estimates.
}
\examples{
bootstrap.correlation(n=2000,cars$speed,cars$dist, plot=TRUE )
}
\author{
Nguyen Khanh Le Ho \cr
}
\keyword{Estimating}
\keyword{and}
\keyword{bootstrap}
\keyword{confidence}
\keyword{correlation,}
\keyword{estimate,}
\keyword{interval}
\keyword{the}

