% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffon.needle.R
\name{buffon.needle}
\alias{buffon.needle}
\title{Simulating Buffon's Needles Experiment}
\usage{
buffon.needle(n=255,l= 2.5, d=3, ind = FALSE, plot= FALSE )
}
\arguments{
\item{n}{the number of needles}

\item{l}{the length of the needles}

\item{d}{the distance between two lines}

\item{ind}{specicfies whether or not to give a red color to the needles that hit the lines}

\item{plot}{specifies whether the board should be drawn.}
}
\value{
An estimation of Pi using the famous buffon's experiment
}
\description{
This function takes as arguments the number of needles \code{n}, the length of needle \code{l} and
the distance between the lines \code{d}, then simulate the buffon needle experiment.
}
\details{
The idea is to generate two random points (x,k) and (xmax, kmax) represents the points which the needle extends.
Any point (x,k) on the board is uniformly distributed, by the assumption that it is equally likely to hit any point on the board.
At the same time an angle theta is generated from U(0, pi/2). Using this angle the point (xmax,kmax) is generated since xmax = x+l*sin(theta)
and kmax = k + l*cos(theta).

The board is designed such that there are 5 parallel lines, distance d from each other i.e we have y_1= d, y_2 = 2*d,...,y_5 = 5*d. Every time there is a hit
it satisfies that kmax > y_i > k.
}
\examples{
buffon.needle(n=2000, l= 2.5 , d= 3 ,  ind = TRUE, plot=TRUE )
}
\author{
Nguyen Khanh Le Ho \cr
}
\keyword{Estimating}
\keyword{buffon}
\keyword{experiment}
\keyword{pi,}
\keyword{the}

