% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.test.R
\name{chi.test}
\alias{chi.test}
\title{Calculating Chi Squared p-value}
\usage{
chi.test(x, p = NULL,plot= FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{p}{a vector of proability, vector of uniform probability if it is unspecified}

\item{plot}{specifies whether or not to make barplot of the observed data and the expectation}
}
\value{
Returns the Chi Squared test statistic.
}
\description{
This function takes as input a numerical vector/obervations \code{x}, and a vector of expected probability \code{p}
then calculate the  \code{Chi Squared Statistic} based on these information.
}
\examples{
chi.test(cars$speed, p = NULL)
}
\author{
Nguyen Khanh Le Ho \cr
}
\keyword{Chi}
\keyword{Statistic}
\keyword{chi}
\keyword{squared}
\keyword{test,}

